+++ utils.h
@@ -2,6 +2,6 @@
 #define UTILS_H
 
 int checkBound(int i);
-double normalDistrib(int x, int y, int sigma);
+double normalDistrib(int x, int y, double sigma);
 
 #endif // UTILS_H
